🧹 Component Refactoring: PropagationControl
Extracts `PropagationInputs` and `ConditionsReadout` subcomponents from `PropagationControl.tsx` to improve maintainability and readability without altering underlying functionality.

*   🎯 **What:** Extracted `PropagationInputs` for the form fields and `ConditionsReadout` for the stats/chart logic into separate files under `src/components/Panel/Propagation/`.
*   💡 **Why:** `PropagationControl.tsx` was over 400 lines long, mixing input state management, condition derivation, and complex rendering. This breaks it down into smaller, logical pieces.
*   ✅ **Verification:** Verified all 531 tests pass (`npm run test`), linting succeeds (`npm run lint`), and visual UI functionality remains identical via Playwright verification scripts.
*   ✨ **Result:** Cleaned up the main `PropagationControl` file to act strictly as an orchestrator, reducing cognitive load when modifying this domain.
